refactor: Move models to suitable paths
